refactor(store): split shared, formulas and folders out of action.ts (#711)
First cut of the 2,929-line file. shared.ts, formulas.ts and folders.ts;
action.ts is 2,769 lines and now the barrel.

Pure moves: every declaration is copied byte-for-byte with the doc
comment above it, so git log --follow keeps working. No consumer import
changes -- action.ts re-exports the new modules, so every existing
`from '../store/databases/action'` resolves as before.

Two changes that are not moves, both forced:

- `log`, `getErrorMsg`, `setDBError` and `clearDBError` gain `export`.
  They were file-local, and they are exactly the symbols reached across
  every concern boundary -- measured, not guessed.
- shared.ts exists at all. Without it every module would depend on the
  one holding those four, which makes the graph a star with a mutable
  hub. With them split out the graph is a tree, and adding a module
  cannot introduce a cycle. Verified: zero cycles across all nine
  planned modules.

Imports were derived per module by scanning the moved text for bindings
the original import block introduces, with comments stripped first so a
mention in prose cannot drag in an import. action.ts's own block was
then pruned the same way -- five bindings dropped.

lint 0 · build 0 · 94 files / 1138 tests, unchanged from before the move
